requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asyncmy==0.2.10
bcrypt==3.2.0
billiard==4.2.1
celery==5.5.0
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
kombu==5.5.2
markdown-it-py==3.0.0
mdurl==0.1.2
passlib==1.7.4
prompt_toolkit==3.0.50
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
rich==14.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
